Implement Conditional Dependencies for Windows
CMakeLists has been modified to ease the entry into creating a new plugin. Now only a single OBSDIR pointing to the root of the obs-studio repository is required to get started, with other dependencies introducing themselves as configurable options. QTDIR is no longer forced, but still cannot be used if it is not set.

The pipelines have not been updated to reflect these changes. Additionally, this probably doesn't work at all on Mac or Linux at the moment. This is considered WIP.

It was suggested that OBS_BUILDDIR be configurable. This is a good idea and will be implemented later.